refactored e2e tests
