+++ git-epic-creator/e2e_tests/src/services/database/__init__.py
@@ -1,0 +1,13 @@
+"""
+Database operations and utilities for e2e tests.
+
+This package provides modular PostgreSQL functionality:
+- validators: Database validation operations
+"""
+
+from .validators import DatabaseValidators
+
+__all__ = [
+    "DatabaseValidators",
+]
+
