+++ git-epic-creator/e2e_tests/src/services/document/__init__.py
@@ -1,0 +1,13 @@
+"""
+Document operations for e2e tests.
+
+This package provides modular document handling functionality:
+- operations: Document upload and processing operations
+"""
+
+from .operations import DocumentOperations
+
+__all__ = [
+    "DocumentOperations",
+]
+
