+++ git-epic-creator/e2e_tests/src/services/workflow/__init__.py
@@ -1,0 +1,16 @@
+"""
+Workflow operations and utilities for e2e tests.
+
+This package provides modular workflow functionality:
+- ui_helpers: UI service SSO/OAuth simulation
+- status_monitor: Project status monitoring
+"""
+
+from .ui_helpers import UIHelpers
+from .status_monitor import StatusMonitor
+
+__all__ = [
+    "UIHelpers",
+    "StatusMonitor",
+]
+
